rag-chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
openai # Keep for now, might be removed later if not needed
python-dotenv
langchain
langchain-openai # Keep for now, might be removed later if not needed
langchain-qdrant
langchain-text-splitters
asyncpg
psycopg2-binary
fastapi-cors
sentence-transformers # For HuggingFaceEmbeddings
langchain-ollama # For Ollama LLM integration